+++ workspace/gh-workflow/src/lib.rs
@@ -1,4 +1,7 @@
 pub mod error;
-pub mod toolchain;
+mod rust_flag;
+mod version;
 pub(crate) mod workflow;
+pub use rust_flag::*;
+pub use version::*;
 pub use workflow::*;
